+++ client_delivery/shared/src/commonMain/kotlin/di/GatewayModule.kt
@@ -3,7 +3,7 @@ package di
 import data.gateway.local.LocalConfigurationGateway
 import data.gateway.remote.FakeRemoteGateway
 import data.gateway.remote.IdentityRemoteGateway
-import domain.gateway.IFakeRemoteGateway
+import domain.gateway.remote.IFakeRemoteGateway
 import domain.gateway.local.ILocalConfigurationGateway
 import domain.gateway.remote.IIdentityRemoteGateway
 import org.koin.core.module.dsl.bind
