+++ client_delivery/shared/src/commonMain/kotlin/domain/gateway/remote/IFakeRemoteGateway.kt
@@ -1,4 +1,4 @@
-package domain.gateway
+package domain.gateway.remote
 
 interface IFakeRemoteGateway {
